requirements.txt
Flask==2.0.0
gunicorn==20.1.0
PyGithub==1.55
pyTelegramBotAPI==3.8.2
